Final Project Rishaan Reddy
In this interactive adventure, players choose to be either James, the clever, or Aaron, the brave. Beginning with 100 health points and an empty inventory, they find themselves in a mysterious cave filled with eerie animal sounds and a distant light. Each decision—whether to chase safety, face a menacing bear, or scavenge on a deserted island—shapes their path. Along the way, they collect weapons and food and encounter obstacles that test their choices. Will they find escape or succumb to the island’s challenges? Every choice counts in this survival tale.